refactor: split create survey class
